Removed config files as discussed offline
