+++ flake.nix
@@ -108,7 +108,7 @@
             });
           });
 
-          propagatedBuildInputs = [ final.solc_0_8_15 final.yul2llvm_cpp final.python310Packages.pyyaml];
+          propagatedBuildInputs = [ final.solc_0_8_15 final.yul2llvm_cpp ];
         }).overridePythonAttrs (old: {
           doCheck = true;
           checkInputs = [
@@ -171,7 +171,6 @@
               # Python component dependencies
               (python3.withPackages (p: [
                 p.venvShellHook
-                p.pyyaml
               ]))
               solc_0_8_15
               jq
